Move config class to new file
